Refactor MainAppBar into modular components
- Split MainAppBar into UserAccountMenu, TopAppBar, NavigationDrawer, and NavigationList components
- Rename MainAppBar to AppLayout to better reflect its responsibility as overall layout manager
- Move components to appropriate folder structure with proper separation of concerns
- Maintain all existing functionality while improving code organization and reusability
